+++ src/templates/defaults.js
@@ -1,3 +1,5 @@
+import { getExtraTemplates } from '../internal';
+
 import storeList from './storeList.hbs';
 import reserve from './reserve.hbs';
 import success from './success.hbs';
@@ -8,4 +10,5 @@ export default {
   reserve,
   success,
   liveInventory,
+  ...getExtraTemplates(),
 };